requirements.txt
python-dotenv   # access keys .env file
pandas
numpy
matplotlib
seaborn
pymongo
certifi
pymongo[srv]
scikit-learn
pyaml
mlflow
dagshub
fastapi
uvicorn
#-e .            # links to `setup.py` file   